+++ dwds/test/instances/record_inspection_test.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+@Tags(['daily'])
 @TestOn('vm')
 @Timeout(Duration(minutes: 2))
 
@@ -263,24 +264,24 @@ Future<void> _runTests({
         expect(await getFields(instanceRef), {
           1: true,
           2: 3,
-          3: {'a': 1, 'b': 5}
+          3: {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 0), {
           1: true,
           2: 3,
-          3: {'a': 1, 'b': 5}
+          3: {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 1), {
           2: 3,
-          3: {'a': 1, 'b': 5}
+          3: {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 1, count: 1), {2: 3});
         expect(await getFields(instanceRef, offset: 1, count: 2), {
           2: 3,
-          3: {'a': 1, 'b': 5}
+          3: {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 2), {
-          3: {'a': 1, 'b': 5}
+          3: {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 3), {});
         expect(await getFields(instanceRef, offset: 0, count: 0), {});
@@ -292,7 +293,7 @@ Future<void> _runTests({
         expect(await getFields(instanceRef, offset: 0, count: 5), {
           1: true,
           2: 3,
-          3: {'a': 1, 'b': 5}
+          3: {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 3, count: 5), {});
       });
@@ -351,24 +352,24 @@ Future<void> _runTests({
         expect(await getFields(instanceRef), {
           1: true,
           2: 3,
-          'array': {'a': 1, 'b': 5}
+          'array': {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 0), {
           1: true,
           2: 3,
-          'array': {'a': 1, 'b': 5}
+          'array': {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 1), {
           2: 3,
-          'array': {'a': 1, 'b': 5}
+          'array': {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 1, count: 1), {2: 3});
         expect(await getFields(instanceRef, offset: 1, count: 2), {
           2: 3,
-          'array': {'a': 1, 'b': 5}
+          'array': {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 2), {
-          'array': {'a': 1, 'b': 5}
+          'array': {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 3), {});
         expect(await getFields(instanceRef, offset: 0, count: 0), {});
@@ -380,7 +381,7 @@ Future<void> _runTests({
         expect(await getFields(instanceRef, offset: 0, count: 5), {
           1: true,
           2: 3,
-          'array': {'a': 1, 'b': 5}
+          'array': {'a': 1, 'b': 5},
         });
         expect(await getFields(instanceRef, offset: 3, count: 5), {});
       });
@@ -438,25 +439,25 @@ Future<void> _runTests({
 
         expect(await getFields(instanceRef), {
           1: true,
-          2: {1: false, 2: 5}
+          2: {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 0), {
           1: true,
-          2: {1: false, 2: 5}
+          2: {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 1), {
-          2: {1: false, 2: 5}
+          2: {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 2), {});
         expect(await getFields(instanceRef, offset: 0, count: 0), {});
         expect(await getFields(instanceRef, offset: 0, count: 1), {1: true});
         expect(await getFields(instanceRef, offset: 0, count: 2), {
           1: true,
-          2: {1: false, 2: 5}
+          2: {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 0, count: 5), {
           1: true,
-          2: {1: false, 2: 5}
+          2: {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 2, count: 5), {});
       });
@@ -509,31 +510,31 @@ Future<void> _runTests({
 
         expect(await getFields(instanceRef), {
           1: true,
-          'inner': {1: false, 2: 5}
+          'inner': {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 0), {
           1: true,
-          'inner': {1: false, 2: 5}
+          'inner': {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 1), {
-          'inner': {1: false, 2: 5}
+          'inner': {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 1, count: 1), {
-          'inner': {1: false, 2: 5}
+          'inner': {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 1, count: 2), {
-          'inner': {1: false, 2: 5}
+          'inner': {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 2), {});
         expect(await getFields(instanceRef, offset: 0, count: 0), {});
         expect(await getFields(instanceRef, offset: 0, count: 1), {1: true});
         expect(await getFields(instanceRef, offset: 0, count: 2), {
           1: true,
-          'inner': {1: false, 2: 5}
+          'inner': {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 0, count: 5), {
           1: true,
-          'inner': {1: false, 2: 5}
+          'inner': {1: false, 2: 5},
         });
         expect(await getFields(instanceRef, offset: 2, count: 5), {});
       });