+++ frontend_server_common/lib/src/asset_server.dart
@@ -17,7 +17,7 @@ import 'package:shelf/shelf.dart' as shelf;
 import 'package:test_common/test_sdk_layout.dart';
 
 class TestAssetServer implements AssetReader {
-  late final String basePath;
+  late final String _basePath;
   final String index;
 
   final _logger = Logger('TestAssetServer');
@@ -43,8 +43,11 @@ class TestAssetServer implements AssetReader {
     this._fileSystem,
     this._sdkLayout,
   ) {
-    basePath = _parseBasePathFromIndexHtml(index);
+    _basePath = _parseBasePathFromIndexHtml(index);
   }
+
+  @override
+  String get basePath => _basePath;
 
   bool hasFile(String path) => _files.containsKey(path);
   Uint8List getFile(String path) => _files[path]!;
@@ -82,7 +85,6 @@ class TestAssetServer implements AssetReader {
       // Assets are served via GET only.
       return shelf.Response.notFound('');
     }
-
     final requestPath = _stripBasePath(request.url.path, basePath);
     if (requestPath == null) {
       return shelf.Response.notFound('');
@@ -90,7 +92,7 @@ class TestAssetServer implements AssetReader {
 
     var headers = <String, String>{};
 
-    if (request.url.path.endsWith('index.html')) {
+    if (request.url.path.endsWith('.html')) {
       var indexFile = _fileSystem.file(index);
       if (indexFile.existsSync()) {
         headers[HttpHeaders.contentTypeHeader] = 'text/html';