+++ webdev/lib/src/serve/webdev_server.dart
@@ -139,6 +139,7 @@ class WebDevServer {
           verbose: options.configuration.verbose,
           experiments: options.configuration.experiments,
           sdkConfigurationProvider: const DefaultSdkConfigurationProvider(),
+          canaryFeatures: options.configuration.canaryFeatures,
         );
       }
       var shouldServeDevTools =