+++ webdev/CHANGELOG.md
@@ -1,4 +1,16 @@
-## 3.0.5-wip
+## 3.0.7-wip
+
+- Update `build_web_compilers` constraint to `^4.0.4`.
+
+## 3.0.6
+
+- Update `dwds` constraint to `19.0.2`.
+
+- Update SDK constraint to `>=3.1.0-254.0.dev <4.0.0`.
+
+## 3.0.5
+
+- Update `dwds` constraint to `19.0.1`.
 
 ## 3.0.4
 
@@ -10,7 +22,7 @@
 
 ## 3.0.2
 
--  Update `build_daemon` constraint to `^4.0.0`.
+- Update `build_daemon` constraint to `^4.0.0`.
 
 ## 3.0.1